+++ 404.php
@@ -1,7 +1,8 @@
 <?php
-require_once('helpers.php');
+require_once('init.php');
+/*require_once('helpers.php');
 require_once('data.php');
-require_once('functions.php');
+require_once('functions.php');*/
 
 /**
  * @var array $categories - массив для вывода категорий
