+++ init.php
@@ -1,0 +1,15 @@
+<?php
+require_once('helpers.php');
+require_once('config.php');
+require_once('functions.php');
+require_once('data.php');
+
+require 'vendor/autoload.php';
+require_once('email-handler.php');
+require_once('getwinner.php');
+require_once('validate-form.php');
+
+
+
+
+
