+++ logout.php
@@ -1,5 +1,6 @@
 <?php
-require_once('config.php');
+require_once('init.php');
+/*require_once('config.php');*/
 
 $_SESSION = [];
 header("Location: /index.php");
